Changed to XSL as code generator

+++ SqlMagic.Host/Program.cs
@@ -1,35 +1,58 @@
-﻿using SqlMagic.CodeGeneration;
-using SqlMagic.Model.Code;
-using SqlMagic.Model.Database;
-using SqlMagic.Output;
+﻿using Flyingpie.Storm.CodeGeneration;
+using Flyingpie.Storm.Model;
+using Flyingpie.Storm.Output;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
 
-namespace SqlMagic.Host
+namespace Flyingpie.Storm.Host
 {
     class Program
     {
         static void Main(string[] args)
         {
-            new Example();
+            Console.WriteLine("Generate or execute? (g/E)");
+            var action = Console.ReadLine();
+
+            if (action.ToUpper() == "G")
+            {
+                Generate();
+            }
+            else
+            {
+                new Example();
+            }
+        }
+
+        static void Generate()
+        {
+            var root = @"C:\Users\marcovdo\Source\Repos\Storm\SqlMagic.Output";
+            var transform = @"Default.xslt";
+            var output = @"Database.Generated.cs";
 
             // Generate database model
+            Console.WriteLine("Generating model from database..");
             var model = new DatabaseModel("Data Source=DBCS_DEV_SQLSRV;Initial catalog=DBCServices5;Integrated Security=true;");
             model.Initialize();
 
-            // Generate code model
-            var code = new CodeModel(model);
-            code.Initialize();
-
             // Generate code
-            var fileName = @"C:\Users\Marco vd Oever\Documents\Visual Studio 2012\Projects\SqlMagic\SqlMagic.Output\Database.Generated.cs";
-            File.Delete(fileName);
-            var stream = File.OpenWrite(fileName);
-            var codeGenerator = new CodeGenerator(code);
-            codeGenerator.Generate(stream);
+            Console.WriteLine("Generating code from model..");
+            var fileName = Path.Combine(root, output);
+            if (File.Exists(fileName))
+            {
+                File.Delete(fileName);
+            }
+
+            //var stream = File.Open(fileName, FileMode.OpenOrCreate);
+            var codeGenerator = new XsltCodeGenerator(model);
+            var code = codeGenerator.Generate(Path.Combine(root, transform));
+
+            var formatter = new DefaultCodeFormatter();
+            code = formatter.Format(code);
+
+            File.WriteAllText(Path.Combine(root, output), code);
         }
     }
 }
